main change: Glasses now use Forge Energy beside of that... fixed some suggestions the IDE made

+++ src/main/java/com/bymarcin/openglasses/utils/OGUtils.java
@@ -49,7 +49,7 @@ public class OGUtils {
 	}
 	
 	public static int getLightLevelPlayer(EntityPlayer e){		
-		return (int) e.world.getLightFor(EnumSkyBlock.SKY, new BlockPos(e.posX, e.posY + 1, e.posZ));
+		return e.world.getLightFor(EnumSkyBlock.SKY, new BlockPos(e.posX, e.posY + 1, e.posZ));
     }
     
     public static boolean isPlayerSwimming(EntityPlayer e){
